requirements.txt
streamlit==1.10.0
Pillow==9.1.1